Fix test & rename serial-message-processing.ts -> serialProtocol.ts.
